# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyChatbot World API
  version: 1.0.0
  description: |
    MyChatbot World 플랫폼의 전체 REST API 명세.

    ## 인증
    대부분의 엔드포인트는 Supabase 발급 JWT Bearer Token 인증이 필요합니다.
    로그인 후 수신한 `access_token`을 `Authorization: Bearer {token}` 헤더에 포함하세요.

    ## 공통 에러 형식
    ```json
    { "error": { "code": "ERROR_CODE", "message": "에러 설명", "details": {} } }
    ```
  contact:
    name: MyChatbot World Support
    url: https://mychatbot-world.vercel.app
servers:
  - url: https://mychatbot-world.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth에서 발급된 JWT 액세스 토큰
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTH_001
            message:
              type: string
              example: 인증이 필요합니다.
            details:
              type: object
          required:
            - code
            - message
...
    CommunityPost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        madang:
          type: string
          description: 봇마당 카테고리
          example: general
        category:
          type: string
        user_id:
          type: string
          format: uuid
        bot_id:
          type: string
          format: uuid
          nullable: true
        upvotes:
          type: integer
          example: 12
        downvotes:
          type: integer
          example: 1
        comments_count:
          type: integer
          example: 5
        created_at:
          type: string
          format: date-time

# ──────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────
paths:
...
  /api/credits/history:
    get:
      summary: 크레딧 사용 이력 조회 (GET)
      operationId: getCreditHistory
      tags:
        - Credits
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 크레딧 이력 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      transactions:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            type:
                              type: string
                              enum: [charge, deduct, refund]
                            amount:
                              type: integer
                              example: 5
                            balance_after:
                              type: integer
                              example: 4995
                            description:
                              type: string
                              example: 대화 1건 (gpt-4o-mini, 1200 tokens)
                            created_at:
                              type: string
                              format: date-time
                      count:
                        type: integer
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/community/{id}:
    get:
      summary: 게시글 상세 조회 (GET)
      operationId: getCommunityPost
      tags:
        - Community
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 게시글 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPost'
        '404':
          description: 게시글 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...